# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlassCrawler API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CrawlRequest {
  location: string;
  productName: string;
  /**
   * Comma-separated list of keywords to exclude
   * @nullable
   */
  negativeKeywords?: string | null;
}

export type CrawlSessionStatus =
  (typeof CrawlSessionStatus)[keyof typeof CrawlSessionStatus];

export const CrawlSessionStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface CrawlSession {
  id: number;
  location: string;
  productName: string;
  /** @nullable */
  negativeKeywords?: string | null;
  status: CrawlSessionStatus;
  pagesLoaded: number;
  itemsFound: number;
  itemsFiltered: number;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Listing {
  id: number;
  sessionId: number;
  olxId: string;
  title: string;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  listingUrl: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  sellerName?: string | null;
  /** @nullable */
  sellerJoinDate?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  listingDate?: string | null;
  createdAt: string;
}

export type WatchlistItemStatus =
  (typeof WatchlistItemStatus)[keyof typeof WatchlistItemStatus];

export const WatchlistItemStatus = {
  active: "active",
  recent: "recent",
} as const;

export interface WatchlistItem {
  id: number;
  olxId: string;
  title: string;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  listingUrl: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  sellerName?: string | null;
  /** @nullable */
  sellerJoinDate?: string | null;
  /** @nullable */
  location?: string | null;
  status: WatchlistItemStatus;
  createdAt: string;
}

export interface AddWatchlistRequest {
  olxId: string;
  title: string;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  listingUrl: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  sellerName?: string | null;
  /** @nullable */
  sellerJoinDate?: string | null;
  /** @nullable */
  location?: string | null;
}

export interface NegativeKeyword {
  id: number;
  keyword: string;
  createdAt: string;
}

export interface AddNegativeKeywordRequest {
  keyword: string;
}

export type GetListingsParams = {
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  sessionId?: number | null;
};