# lib/api-zod/src/generated/types/crawlSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlassCrawler API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CrawlSessionStatus } from "./crawlSessionStatus";

export interface CrawlSession {
  id: number;
  location: string;
  productName: string;
  /** @nullable */
  negativeKeywords?: string | null;
  status: CrawlSessionStatus;
  pagesLoaded: number;
  itemsFound: number;
  itemsFiltered: number;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: Date;
  updatedAt: Date;
}