# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlassCrawler API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Listing {
  id: number;
  sessionId: number;
  olxId: string;
  title: string;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  listingUrl: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  sellerName?: string | null;
  /** @nullable */
  sellerJoinDate?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  listingDate?: string | null;
  createdAt: Date;
}