# lib/api-zod/src/generated/types/negativeKeyword.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlassCrawler API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NegativeKeyword {
  id: number;
  keyword: string;
  createdAt: Date;
}